Use nulls, not empty strings, to mark missing things in JSON
Before this commit, if something (e.g. a resource description) was
missing in the blueprint, the value of a corresponding property in the
AST was |null|. This was not the case for JSON representation however --
the value was converted to an empty string there.

This inconsistency was present only for compatibility with an older
parser used internally by Apiary. This commit removes it so that users
are not burdened by it.